+++ src/dockings/grid_representation.jl
@@ -5,7 +5,7 @@ include("create_centroids.jl")
 include("set_marked_cells.jl")
 include("create_inner_outer_grid.jl")
 
-function grid_representation(atoms::Vector{Vector3{Float32}}, gridsize::Int64, centroids::Array{Meshes.Point3, 3})
+function grid_representation(atoms::Vector{Vector3{Float32}}, gridsize::Int32, centroids::Array{Meshes.Point3, 3})
     # calculate atomballs around proteins atoms
     atomballs = create_atomballs(atoms)
     # find all cells in grid which represent protein
